Refactor out EnvironmentV1.
